requirements.txt
flask
mediapipe
opencv-python
numpy
google-api-python-client
boto3
gunicorn
